% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-marray-as.array-as.vector.r,
%   R/method-mmatrix-as.matrix-as.vector.r, R/method-mvector-as.vector.r
\name{as.array.marray}
\alias{as.array.marray}
\alias{as.vector.marray}
\alias{as.matrix.mmatrix}
\alias{as.vector.mmatrix}
\alias{as.vector.mvector}
\title{Converting memory-mapped objects to R objects}
\usage{
\method{as.array}{marray}(x, ...)

\method{as.vector}{marray}(x, mode = "any")

\method{as.matrix}{mmatrix}(x, ...)

\method{as.vector}{mmatrix}(x, mode = "any")

\method{as.vector}{mvector}(x, mode = "any")
}
\arguments{
\item{x}{memory-mapped object to convert}

\item{...}{extra parameters (ignored)}

\item{mode}{the mode oh the created vector}
}
\description{
Converting memory-mapped objects to R objects
}
