% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-marray-as.array-as.vector.r
\name{as.array.marray}
\alias{as.array.marray}
\title{Converting memory mapped vectors (on disk) 
to R vectors (in memory)}
\usage{
\method{as.array}{marray}(x)
}
\description{
Converting memory mapped vectors (on disk) 
to R vectors (in memory)
}
