% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-mmatrix-as.matrix.r
\name{as.matrix.mmatrix}
\alias{as.matrix.mmatrix}
\title{Converting memory mapped matrices (on disk) 
to R matrices (in memory)}
\usage{
\method{as.matrix}{mmatrix}(x)
}
\description{
Converting memory mapped matrices (on disk) 
to R matrices (in memory)
}
