% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-mvector-as.vector.r
\name{as.vector.mvector}
\alias{as.vector.mvector}
\title{Converting memory mapped vectors (on disk) 
to R vectors (in memory)}
\usage{
\method{as.vector}{mvector}(x, mode = "any")
}
\description{
Converting memory mapped vectors (on disk) 
to R vectors (in memory)
}
