% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy.values.r
\name{copy.values}
\alias{copy.values}
\alias{copy.values,memoryMapped,numericOrArray-method}
\alias{copy.values,memoryMapped,memoryMapped-method}
\title{Copy values to memory mapped object}
\usage{
copy.values(x, values)

\S4method{copy.values}{memoryMapped,numericOrArray}(x, values)

\S4method{copy.values}{memoryMapped,memoryMapped}(x, values)
}
\arguments{
\item{x}{a memory mapped object}

\item{values}{a R object or a memory mmaped object}
}
\description{
Copy values to memory mapped object
}
\details{
Copy \code{values} to \code{x}, recycling if necessary. This
function modifies \code{x} in-place.
}
