% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-mmatrix-add.descriptor.file.r
\name{add.descriptor.file}
\alias{add.descriptor.file}
\alias{add.descriptor.file,mmatrix-method}
\alias{add.descriptor.file,mvector-method}
\title{Descriptor file}
\usage{
add.descriptor.file(object)

\S4method{add.descriptor.file}{mmatrix}(object)

\S4method{add.descriptor.file}{mvector}(object)
}
\arguments{
\item{object}{a mmatrix or mvector object}
}
\description{
Descriptor file
}
\details{
Creates a descriptor file, for reading the matrix with the package bigmemory or the \link{read.descriptor} function.
}
