% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mvector.r
\docType{class}
\name{mvector-class}
\alias{mvector-class}
\title{Class \code{mvector}}
\description{
A class for manipulation of memory mapped file
}
\section{Slots}{

\describe{
\item{\code{ptr}}{\code{externalptr} to an instance of the C++ \code{MMatrix} class, itself handling
the internal memory structure of the file.}

\item{\code{file}}{\code{character} with the path (absolute) of the file used to store the mmatrix.}

\item{\code{length}}{\code{integer} the length of the vector}

\item{\code{datatype}}{\code{character} with the C++ underlying datatype.}

\item{\code{readonly}}{\code{logical} Indicates if the matrix if read only.}
}}

\seealso{
\link{marray-class}, \link{mmatrix-class}
}
