% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.descriptor.r
\name{read.descriptor}
\alias{read.descriptor}
\title{Read big memory descriptor file}
\usage{
read.descriptor(basename)
}
\arguments{
\item{basename}{basename of the file}
}
\description{
Read big memory descriptor file
}
\details{
Creates a mmatrix by reading a 'bigmemory' descriptor file.
The descriptor file name is obtained by appending ".desc" to the basename.
}
\seealso{
\link{add.descriptor.file}
}
