% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/houba.r
\name{houba}
\alias{houba}
\title{Set maxsize for R objects}
\usage{
houba(...)
}
\arguments{
\item{max.size}{maximum size of a R object created by subsetting an mvector or an mmatrix}
}
\description{
Set maxsize for R objects
}
\details{
when subsetting an mvector or an mmatrix, if the size of the resulting object 
is greater than `max.size`, then the result will be a memory mapped object (mvector or mmatrix),
else if will be a R object (vector or matrix).
The default value is 1e6.
Set `max.size` to `0` to always get a memory mapped object and to `Inf` to always get a R object.
}
