% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-marray.r
\docType{class}
\name{marray-class}
\alias{marray-class}
\title{Class \code{"marray"}}
\description{
S4 class for memory-mapped arrays
}
\section{Slots}{

\describe{
\item{\code{ptr}}{\code{externalptr} to an instance of the C++ \code{MMatrix} class, itself handling
the internal memory structure of the file.}

\item{\code{file}}{\code{character} with the path (absolute) of the file used to store the mmatrix.}

\item{\code{dim}}{\code{vector} keeping the dimensions of the matrix. Using \code{ncol(x = mmatrix)} or \code{nrow(x = mmatrix)}
will respectively link to \code{dim[1]} and \code{dim[2]}
because the size of mapping depends on this dim, changing it should not be possible nor effective}

\item{\code{datatype}}{\code{character} with the C++ underlying datatype.}

\item{\code{readonly}}{\code{logical} Indicates if the matrix if read only.}
}}

\section{Objects from the Class}{

Objects can be created by calls of the form \code{marray(...)}.
}

\seealso{
\link{mmatrix-class}, \link{mvector-class}

\link{read.mmatrix}, \link{as.matrix.mmatrix}, \link{create.descriptor.file}
}
