% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mmatrix.r, R/method-subsetting.r
\docType{class}
\name{mmatrix-class}
\alias{mmatrix-class}
\alias{[,mmatrix,numeric,numeric-method}
\alias{[,mmatrix,numeric,missing-method}
\alias{[,mmatrix,missing,missing-method}
\alias{[,mvector,numeric,missing-method}
\alias{[,mvector,missing,missing-method}
\title{Class \code{"mmatrix"}}
\usage{
\S4method{[}{mmatrix,numeric,numeric}(x, i, j, ..., drop = TRUE)

\S4method{[}{mmatrix,numeric,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{mmatrix,missing,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{mvector,numeric,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{mvector,missing,missing}(x, i, j, ..., drop = TRUE)
}
\description{
S4 class for memory-mapped matrices of type \code{double}, \code{float}, \code{int} and \code{short}
}
\details{
If you wish to work with a matrix with a \code{ncol == 1}, try looking into
the \link{mvector-class}.
}
\section{Slots}{

\describe{
\item{\code{ptr}}{\code{externalptr} to an instance of the C++ \code{MMatrix} class, itself handling
the internal memory structure of the file.}

\item{\code{file}}{\code{character} with the path (absolute) of the file used to store the mmatrix.}

\item{\code{dim}}{\code{vector} keeping the dimensions of the matrix. Using \code{ncol(x = mmatrix)} or \code{nrow(x = mmatrix)}
will respectively link to \code{dim[1]} and \code{dim[2]}
because the size of mapping depends on this dim, changing it should not be possible nor effective}

\item{\code{datatype}}{\code{character} with the C++ underlying datatype.}
}}

\section{Objects from the Class}{

Objects can be created by calls of the form \code{mmatrix(...)}, using the \link{mmatrix} function.
}

\section{Methods}{

\describe{
\item{[}{\code{signature(x = "mmatrix", i = "numeric" or "logical" or "missing",}\cr\code{j = "numeric" or "logical" or "missing", drop = "missing")}:
\cr Extract a sub-matrix (a new \code{mmatrix}). }

\item{as.matrix}{\code{signature(x = "mmatrix")}:
\cr Convert a \code{mmatrix} into a \link{matrix}.
The whole point of this package was to NOT have an enormous matrix object,
but you can use this method for a small mmatrix or a snippet !}

\item{add.descriptor.file}{\code{signature(x = "mmatrix")}:
\cr create a .desc file linked to the mmatrix. 
Having this file coupled with the mmatrix file allows you 
to re-open the mmatrix elsewhere, and works essentially
as a keep for all the information present in the slots.
This is compatible with the package \code{big.memory} and its function \code{attach.big.matrix()}
}

\item{read.mmatrix}{\code{signature(x = "basename")}:
\cr Take a descriptor.file and use it to open an mmatrix.
}

\item{restore}{\code{signature(x = "mmatrix")}:
\cr Regain access to the \code{mmatrix} content when the \code{ptr} to it is broken (null)
by remapping the file.
}
}
}

\examples{

showClass("mmatrix")
}
\seealso{
\link{read.mmatrix}, \link{as.matrix.mmatrix}, \link{create.descriptor.file}, \link{restore.mmatrix}
}
