% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.descriptor.r
\name{read.descriptor}
\alias{read.descriptor}
\title{Read big memory descriptor file}
\usage{
read.descriptor(basename, readonly)
}
\arguments{
\item{basename}{basename of the file}

\item{readonly}{\code{TRUE} by default, specifies if the object should be readonly}
}
\description{
Read big memory descriptor file
}
\details{
Creates a \code{mmatrix} or a \code{mvector} by reading a 'bigmemory' descriptor file.
The descriptor file name is obtained by appending ".desc" to the basename.
}
\seealso{
\link{add.descriptor.file}
}
